archive/requirements.txt
flask==2.3.3
numpy==1.24.3
scikit-learn==1.5.0
pandas==2.0.3
joblib==1.3.1
prometheus-client==0.17.1
requests==2.31.0
google-cloud-bigquery==3.26.0
scipy==1.11.4
typing-extensions==4.9.0
sqlalchemy==2.0.25
db-dtypes==1.2.0
